Bugfix + Features + Root implementations
- Bugfix in PGridListLayout. A null constraint was not correctly
converted to a valid PTableCellIndex.
- Added methods to generate a copy of PColor with multiplied alpha
- Added methods to PRenderer to draw a String as a char array
- Changed signature of getLayoutPreference to return a PLayoutPreference
instead of a DefaultPLayoutPreference. The previous version was a bug.
- Updated work on the LWJGL3 renderers. Still a WIP and far from
finished.

+++ piq-lwjgl3/src/edu/udo/piq/lwjgl3/renderer/LwjglRmXor.java
@@ -54,7 +54,7 @@ public class LwjglRmXor extends LwjglRmOutline implements LwjglRenderMode {
 	}
 	
 	@Override
-	public void drawEllipseImmediate(int x, int y, int width, int height) {
+	public void drawEllipseImmediate(float x, float y, float width, float height) {
 		GL11.glEnable(GL11.GL_COLOR_LOGIC_OP);
 		GL11.glLogicOp(GL11.GL_XOR);
 		super.drawEllipseImmediate(x, y, width, height);
